fix(travis): bumps the minimum version of rust to 1.17.0
This is due in part due to dependencies having evolved (notably libc) and requiring a newer version but also the fix for the clippy linting bumped it up as well. I don't see any major issues with this as the new minimum version is over a year old and I hope no one is still using it, let along using it and this project at the same time.