Fix the AttachOptions.part_scan function
The function currently writes a value of 16 to lo_flags, but 16 is the value for
LO_FLAGS_DIRECT_IO.  The intended flag, LO_FLAGS_PARTSCAN, has a value of 8 [0].

It looks like this was a regression when deprecating attach_with_partscan[1],
which had the correct value of 8.  That same commit introduces this bug.

[0] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/linux/loop.h?h=v5.15#n25
[1] https://github.com/mdaffin/loopdev/commit/1ec98260129771fdb14b064008d857f72e2d5d16