Add github actions workflow
Check compilation and lints for the targets

* aarch64-linux-android
* aarch64-unknown-linux-gnu
* aarch64-unknown-linux-musl
* x86_64-unknown-linux-gnu
* x86_64-unknown-linux-musl

The loopdev crates makes use of bindgen generated code which uses system
headers that can be different for different platforms. Ensure that at
it at least builds.

Enabling the tests is hard with the current structure of the integration
tests and safed for a later PR.