Fixes the add_a_loop_device test
It tried to count the existing devices by looking at the /dev/loop* files. But these may or may not be free devices so the check for an error on existing device file was failing as the count it came up with could be added. It has been changed to try adding the device twice - the first should succeed and the second should fail. This way around we garuntee the state of the first (assuming it was not added outside of the tests).